Refactored implementation source files so that they are better organized when we start adding more implementation types.
