+++ MuyGPyS/_src/optimize/chassis/jax.py
@@ -8,7 +8,7 @@ from typing import Callable
 from MuyGPyS import config, jax_config
 
 from MuyGPyS.gp.muygps import MuyGPS
-from MuyGPyS._src.optimize.numpy_chassis import (
+from MuyGPyS._src.optimize.chassis.numpy import (
     _scipy_optimize as _numpy_scipy_optimize,
     _bayes_opt_optimize as _numpy_bayes_opt_optimize,
 )
