+++ MuyGPyS/_src/optimize/objective/__init__.py
@@ -1,0 +1,4 @@
+# Copyright 2021-2022 Lawrence Livermore National Security, LLC and other
+# MuyGPyS Project Developers. See the top-level COPYRIGHT file for details.
+#
+# SPDX-License-Identifier: MIT
