+++ MuyGPyS/_src/optimize/objective/jax.py
