+++ MuyGPyS/_src/optimize/objective/numpy.py
