+++ MuyGPyS/gp/distance.py
@@ -80,14 +80,14 @@ from typing import Optional, Tuple
 from MuyGPyS import config
 
 if config.muygpys_jax_enabled is False:  # type: ignore
-    from MuyGPyS._src.gp.numpy_distance import (
+    from MuyGPyS._src.gp.distance.numpy import (
         _make_regress_tensors,
         _make_train_tensors,
         _crosswise_distances,
         _pairwise_distances,
     )
 else:
-    from MuyGPyS._src.gp.jax_distance import (
+    from MuyGPyS._src.gp.distance.jax import (
         _make_regress_tensors,
         _make_train_tensors,
         _crosswise_distances,
