+++ MuyGPyS/gp/kernels.py
@@ -49,7 +49,7 @@ from typing import cast, Callable, Dict, List, Optional, Tuple, Union
 from MuyGPyS import config
 
 if config.muygpys_jax_enabled is False:  # type: ignore
-    from MuyGPyS._src.gp.numpy_kernels import (
+    from MuyGPyS._src.gp.kernels.numpy import (
         _rbf_fn,
         _matern_05_fn,
         _matern_15_fn,
@@ -58,7 +58,7 @@ if config.muygpys_jax_enabled is False:  # type: ignore
         _matern_gen_fn,
     )
 else:
-    from MuyGPyS._src.gp.jax_kernels import (
+    from MuyGPyS._src.gp.kernels.jax import (
         _rbf_fn,
         _matern_05_fn,
         _matern_15_fn,
