+++ MuyGPyS/optimize/objective.py
@@ -17,9 +17,9 @@ from typing import Callable
 from MuyGPyS import config
 
 if config.muygpys_jax_enabled is False:  # type: ignore
-    from MuyGPyS._src.optimize.numpy_objective import _mse_fn, _cross_entropy_fn
+    from MuyGPyS._src.optimize.objective.numpy import _mse_fn, _cross_entropy_fn
 else:
-    from MuyGPyS._src.optimize.jax_objective import _mse_fn, _cross_entropy_fn
+    from MuyGPyS._src.optimize.objective.jax import _mse_fn, _cross_entropy_fn
 
 
 def get_loss_func(loss_method: str) -> Callable:
